+++ HISTORY.rst
@@ -28,6 +28,7 @@ Release History
 * Feature: ConfigScopes now extract docstrings and line comments and display
            them when calling ``print_config``
 * Feature: observers are now run in order of priority (settable)
+* Feature: new ``--name=NAME`` option to set the name of experiment for this run
 * API Change: all times are now in UTC
 * API Change: significantly changed the mongoDB layout
 * API Change: MongoObserver and FileStorageObserver now use consecutive